Version 5: Using two modules to separate the main module from the class definition module.
