+++ HelloMessageModule.py
@@ -1,0 +1,10 @@
+class HelloMessage:
+    def show(self):
+        helloMsg = 'Hello Programming World'
+        print(helloMsg)
+        
+        courseMsg = 'Programming Principles with Python'
+        print(courseMsg)
+        
+        courseNo = 10004
+        print('#prog', courseNo, sep='')
